Move PaperMetadata to o.a.s.metrics so that it doesn't clash with s2 PaperMetadata
