+++ src/main/scala/org/allenai/scholar/metrics/metadata/MetadataErrorAnalysis.scala
@@ -2,7 +2,7 @@ package org.allenai.scholar.metrics.metadata
 
 import java.time.Year
 
-import org.allenai.scholar.metrics.ErrorAnalysis
+import org.allenai.scholar.metrics.{ PaperMetadata, ErrorAnalysis }
 import org.allenai.scholar._
 
 object MetadataErrorAnalysis {
